+++ src/main/java/lotto/LottoGame.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.domain.Lotto;
+import lotto.domain.Lottos;
+import lotto.domain.PurchaseAmount;
 import lotto.view.InputView;
 import lotto.view.ResultView;
