+++ src/main/java/lotto/LottoGenerator.java
@@ -1,6 +1,8 @@
 package lotto;
 
 
+import lotto.domain.Lotto;
+
 public interface LottoGenerator {
 
   Lotto generate();
