+++ src/main/java/lotto/LottoReport.java
@@ -1,5 +1,10 @@
 package lotto;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoPrize;
+import lotto.domain.Lottos;
+import lotto.domain.PurchaseAmount;
+
 import java.util.ArrayList;
 import java.util.List;
