+++ src/main/java/lotto/RandomLottoGenerator.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
