+++ src/main/java/lotto/domain/LottoPrize.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.Set;
